Decouple ModelTests
